test(wrap): split seccomp integration coverage by platform
